% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicationTiming-addReplicationTimingsignal.R
\name{addReplicationTimingSignal}
\alias{addReplicationTimingSignal}
\title{Add replication timing signal value}
\usage{
addReplicationTimingSignal(
  reducedFunseqOutput,
  replicationTimingDF,
  useCores = 1
)
}
\arguments{
\item{reducedFunseqOutput}{reducedFunseqOutput data frame}

\item{replicationTimingDF}{replicationTimingDF data frame}

\item{useCores}{Default is one, number of cpu to use}
}
\value{
reducedFunseqOutput data frame
}
\description{
Add replication timing signal value
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{CNCDriver}
