% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somaticSignature-addTriNucleotideDistribution.R
\name{addTriNucleotideDistribution}
\alias{addTriNucleotideDistribution}
\title{Add trinucleotide distribution annotations}
\usage{
addTriNucleotideDistribution(dat)
}
\arguments{
\item{dat}{reducedFunseqOutput data frame}
}
\value{
triNucleotideDat data frame
}
\description{
Add trinucleotide distribution annotations
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
