% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils-calculatePvalueWithMutCategoryDistributionMatched7.R
\name{calculatePvalueWithMutCategoryDistributionMatched7}
\alias{calculatePvalueWithMutCategoryDistributionMatched7}
\title{calculate Pvalue With Mut CategoryDistribution Matched}
\usage{
calculatePvalueWithMutCategoryDistributionMatched7(variantTriMutCategory,
  backgroundPosVariant, mutationDistMatrix, featureDF, reSampleNum, replaceFlag,
  fileName, debugMode = FALSE)
}
\arguments{
\item{variantTriMutCategory}{geneDFunique data frame}

\item{backgroundPosVariant}{mutationDistMatrix data frame}

\item{mutationDistMatrix}{default is one}

\item{featureDF}{geneDFunique data frame}

\item{reSampleNum}{mutationDistMatrix data frame}

\item{replaceFlag}{default is one}

\item{fileName}{fileName data frame}

\item{debugMode}{default is FALSE}
}
\value{
variantTriMutCategoryParsed data frame
}
\description{
calculate Pvalue With Mut CategoryDistribution Matched
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
