% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-extractElementMutations.R
\name{extractElementMutations}
\alias{extractElementMutations}
\title{Extract overlapping mutations from element annotation bed file}
\usage{
extractElementMutations(elementBedfile, reducedFunseqOutput,
  debugMode = FALSE)
}
\arguments{
\item{elementBedfile}{The data frame for element annotated bed file}

\item{reducedFunseqOutput}{reducedFunseq2 data frame}

\item{debugMode}{TRUE or FALSE}
}
\value{
reducedFunseq2 data frame
}
\description{
Extract overlapping mutations from element annotation bed file
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
