% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/replicationTiming-makeElementReplicationTiming.R
\name{makeElementReplicationTiming}
\alias{makeElementReplicationTiming}
\title{Make replication timing signal value for each regulatory element}
\usage{
makeElementReplicationTiming(replicationTimingDF, elementDF, useCores = 1)
}
\arguments{
\item{replicationTimingDF}{replicationTimingDF data frame}

\item{elementDF}{regulatory element data frame}

\item{useCores}{Default is one, number of cpu to use}
}
\value{
elementsRTdf data frame
}
\description{
Make replication timing signal value for each regulatory element
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{CNCDriver}
