% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-makeQQplot.R
\name{makeQQplot}
\alias{makeQQplot}
\title{Make QQ-plot}
\usage{
makeQQplot(outputDf)
}
\arguments{
\item{outputDf}{Output from CNCDriver p-value calculations}
}
\value{
ggplot2 graph object
}
\description{
Make QQ-plot
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{CNCDriver}
