% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mergeVariantPosition_v7.R
\name{mergeVariantPosition}
\alias{mergeVariantPosition}
\title{mergeVariantPosition}
\usage{
mergeVariantPosition(datDF, sizeFactor, countsCutOff = 2)
}
\arguments{
\item{sizeFactor}{sizeFactor data frame}

\item{countsCutOff}{default is two}

\item{geneDF}{list of geneDF data frame}
}
\value{
collapsedDF data frame
}
\description{
mergeVariantPosition
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
