% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils-mergeVariantPosition_v7_enhancerUnit_Cluster.R
\name{mergeVariantPositionClusterEnhancerUnitV2}
\alias{mergeVariantPositionClusterEnhancerUnitV2}
\title{mergeVariantPositionClusterEnhancerUnitV2}
\usage{
mergeVariantPositionClusterEnhancerUnitV2(datDF, sizeFactor, countsCutOff = 2)
}
\arguments{
\item{datDF}{list of geneDF data frame}

\item{sizeFactor}{sizeFactor vector, a parameter only used in pancancer calculation. A sizeFactor for total mutations counts normalization among multiple cancer types}

\item{countsCutOff}{default is two, a parameter only used in pancancer calculation. Scale the mutation occurence value when the counts is equal or above countsCutOff}
}
\value{
collapsedDF data frame
}
\description{
mergeVariantPositionClusterEnhancerUnitV2
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
