% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-mergeVariantPositionSimple.R
\name{mergeVariantPositionSimple}
\alias{mergeVariantPositionSimple}
\title{Calculate recurrence number}
\usage{
mergeVariantPositionSimple(datDF)
}
\arguments{
\item{datDF}{geneDFunique data frame}
}
\value{
results data frame
}
\description{
Calculate recurrence number
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{CNCDriver}
