% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parseMOTIFBR.R
\name{parseMOTIFBR}
\alias{parseMOTIFBR}
\title{Parse MOTIFBR annotations from FunSeq2}
\usage{
parseMOTIFBR(dat, useCores = 1)
}
\arguments{
\item{dat}{reducedFunseqOutput data frame}

\item{useCores}{default is one}
}
\value{
motifBreakParsed data frame
}
\description{
Parse MOTIFBR annotations from FunSeq2
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
