% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-parseMOTIFG.R
\name{parseMOTIFG}
\alias{parseMOTIFG}
\title{Parse MOTIFBR annotations from FunSeq2}
\usage{
parseMOTIFG(dat, useCores = 1)
}
\arguments{
\item{dat}{reducedFunseqOutput data frame}

\item{useCores}{default is one}
}
\value{
motifGainParsed data frame
}
\description{
Parse MOTIFBR annotations from FunSeq2
}
\examples{
# TCF12_disc4[chr19:49990691-49990691#2.23] 
# motifmodel # mutation position # motif start # motif end # motif direction # mutation position on the motif model # delta score 
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
