% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils-parsePosVariantByTriMutcontextWithAnnotationCDS.R
\name{parsePosVariantByTriMutContextWithAnnotationCDS}
\alias{parsePosVariantByTriMutContextWithAnnotationCDS}
\title{Parse Pos Variant By TriMutContext With Annotation for protein coding region}
\usage{
parsePosVariantByTriMutContextWithAnnotationCDS(geneDFunique,
  mutationDistMatrix, useCores = 1)
}
\arguments{
\item{geneDFunique}{geneDFunique data frame}

\item{mutationDistMatrix}{mutationDistMatrix data frame}

\item{useCores}{default is one, number of cput to use}
}
\value{
variantTriMutCategoryParsed data frame
}
\description{
Parse Pos Variant By TriMutContext With Annotation for protein coding region
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{
CNCDriver
}
