% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-preProcessVCF.R
\name{preProcessVCF}
\alias{preProcessVCF}
\title{Parse Funseq2 VCF file}
\usage{
preProcessVCF(
  funseq2OutputFile,
  outputDir,
  tumorType,
  useCores = 6,
  debugMode = FALSE
)
}
\arguments{
\item{funseq2OutputFile}{Where is annotated funseq2 VCF file}

\item{outputDir}{path to store processed Rd file}

\item{tumorType}{study name}

\item{useCores}{number of cores to use for parallelization}

\item{debugMode}{TRUE or FALSE}
}
\value{
flag good or bad
}
\description{
Parse Funseq2 VCF file
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{CNCDriver}
