% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{reducedFunseqOutputCDS}
\alias{reducedFunseqOutputCDS}
\title{reduced funseq2 CDS output for LGG}
\format{A data frame with 18 rows and 3 variables}
\usage{
reducedFunseqOutputCDS
}
\description{
This data contains the topology of a 
circuit with ten genes A1...A5 and B1...B5.
}
\keyword{datasets}
