% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/utils-calculatePvalueWithMutCategoryDistributionMatched7.R
\name{calculatePvalueWithMutCategoryDistributionMatched7}
\alias{calculatePvalueWithMutCategoryDistributionMatched7}
\title{calculate Pvalue With Mut CategoryDistribution Matched}
\usage{
calculatePvalueWithMutCategoryDistributionMatched7(variantTriMutCategory,
  backgroundPosVariant, mutationDistMatrix, featureDF, reSampleNum,
  replaceFlag = FALSE, replicationTimingCutOff, debugFileName,
  debugMode = FALSE)
}
\arguments{
\item{variantTriMutCategory}{variantTriMutCategory data frame}

\item{backgroundPosVariant}{backgroundPosVariant data frame}

\item{mutationDistMatrix}{mutationDistMatrix data frame}

\item{featureDF}{featureDF data frame}

\item{reSampleNum}{number of re-sampling iterations}

\item{replaceFlag}{default is FALSE, whether to replace or not in the sampling process}

\item{replicationTimingCutOff, }{Default is 0.2, a numeric value ranging from 0 to 1}

\item{debugFileName}{fileName to output debug information}

\item{debugMode}{default is FALSE, TRUE or FALSE}
}
\value{
variantTriMutCategoryParsed data frame
}
\description{
calculate Pvalue With Mut CategoryDistribution Matched
}
\examples{
#date<-getRunDates(latest=TRUE)
cancerType<-"KIRC"
selectedSampleId<-NA
#worDir<-getwd()
mutSig2CVthreshold<-0.1
rareMutationUpperLimit<-0.3
rareMutationLowerLimit<-0.1
rareMutationFreq<-0.02

#runNetBox2(dataDir,cancerType,
#           mutationList,ampGeneList,delGeneList,epiSilencedList,
#           mutationFreq,ampGeneFreq,delGeneFreq,epiSilencedFreq,
#           pathwayCommonsDb,directed,
#           linkerPValThreshold,communityDetectionMethod,
#           keepIsolatedNodes,verbose=TRUE)

}
\concept{CNCDriver}
